+++ MaterialDesignExtensions/Controls/TextBoxSuggestions.cs
@@ -6,7 +6,6 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Controls.Primitives;
 using System.Windows.Input;
 using System.Windows.Markup;
 
@@ -166,7 +165,7 @@ namespace MaterialDesignExtensions.Controls
         {
             if (TextBox != null)
             {
-                TextBox.Text = args.Parameter as string ?? "";
+                TextBox.Text = args.Parameter as string ?? string.Empty;
 
                 if (KeepFocusOnSelection)
                 {